Fix objects showing through closer surfaces
z_perspective is true seems to imply w-buffering and then the w-coordinate
stored in the depth buffer should also be interpolated in a
perspective-correct way. We do this by calculating w and setting
gl_FragDepth in the fragment shader.

Since enabling polygon offset and setting values using glPolygonOffset
won't have any effect when manually setting gl_FragDepth, we introduce the
depthOffset variable to obtain similar behaviour (but the glPolygonOffset
factor-argument is currently not emulated.)

We also change custom clip plane behaviour (i.e. gl_ClipDistance[i]),
which seemed wrong, but not sure if the present fix is correct either.
Note that in Halo 2 the model-view-projection matrix has the projection
matrix setup with near plane approx. 256.0 and far plane 16777215.0.
Coincidentally, the NV_PGRAPH_ZCLIPMIN and NV_PGRAPH_ZCLIPMAX registers
also contain approx. 256.0 and 16777215.0. Therefore it doesn't make sense
to use the z-coordinate in gl_ClipDistance calculations, because the
z-coordinate has already been transformed to the clipping range by the
prior calculations. When w-buffering is not used, I'm not sure if clipping
by w makes sense or what the meaning of NV_PGRAPH_ZCLIPMIN and
NV_PGRAPH_ZCLIPMAX are.

(Note: in Midtown Madness 3 the clip values are approx. 4417.0 and
16777215.0, so this setup is not just a Halo 2 fluke.)

Halo 2 renders some HUD elements using z=0.0 and w=1.0. These have
glDisable(GL_DEPTH_TEST). We disable clip planes when depth testing is
disabled, because 1.0 < 256.0 and the HUD elements would otherwise get
clipped away.

There is a subtle floating point accuracy issue in the vertex shader when
converting the z-coordinate from the Xbox vertex shader output to the
OpenGL range. Halo 2 Xbox vertex shaders output some HUD elements with
z=16777215.0, w=16777215.0. We convert it using
"oPos.z = (2.0*oPos.z - clipRange.y)/clipRange.y;"
and after multiplication by w, we get 16777215.0 back. If this was changed
to the equivalent (in exact arithmetic)
"oPos.z = 2.0*oPos.z/clipRange.y - 1.0;"
we would get 16777216.0 after multiplying by w=16777215.0 and then the HUD
elements would be clipped away. The source of the inaccuracy is likely due
to the division by clipRange.y being done by multiplying by the reciprocal
of clipRange.y under the hood.

We also multiply by w in the vertex shader, because we need to be in
clip-space coordinates for gl_ClipDistance to have a chance of working
properly.