Hack for Halo 2 to fix shadow flickering
Halo 2 Beaver Creek map has small rocks with shadows which can have
flickering horizontal lines. The shadows are drawn on the ground in
another pass with the same models as for the ground, but for some reason
with glEnable(GL_DEPTH_CLAMP). The flickering may be due to small
differences in the coordinates since the usual OpenGL clipping to
-w <= z <= w is now disabled compared to be when rendering the ground. The
shadows are drawn with depth function GL_EQUAL so there is no tolerance
for any differences.

We fix this by a hack to disable the depth clamping, but the detection is
rather crude and may cause visual issues elsewhere or in other games. Why
is depth clamping enabled here anyway? Is it enabled due to some Xemu
emulation bug?