Fixes for signed and unsigned texture channels and bump environment map
Introduces unsigned integer texture stage RGBA result. This is hinted in
https://registry.khronos.org/OpenGL/extensions/NV/NV_texture_shader.txt.
According to that document, texture stages produce two separate results:
a shader stage result (it%d, dot%d) that can be used in subsequent texture
stages, and a texture unit RGBA result (t%d) that is used in register
combiners. There also may be a HILO value shader stage result, but I'm not
sure if it exists in Xbox nv2a and this commit does not implement such a
thing.

This also adds handling for the 16-bit texture R16B16 (0x33). It would
make sense if it could be used with BUMPENVMAP mode, but I was unable to
to make a real Xbox apply it as expected. Perhaps Xbox can't. After
sampling e.g. R16B16, Xbox seems to create a 32-bit integer BGRA value
from the two 16-bit channels in the texture and sets it as the shader
stage result. That's what this commit also emulates.

Much of the above Xbox signed/unsigned and integer rounding behaviour can
be tested using the new BumpMapTests and BumpEnvLumTests in
nxdk_pgraph_tests repository.

Comment on the game Jet Set Radio Future: the boost dash effect in the
game seems to have a small bug. It uses the BUMPENVMAP mode with an 8-bits
per channel RGB texture for displacement, but appears to assume that the
displacement channels are R and G, unsigned (since the texture data implies
that 0x80 should mean no displacement.) However, Xbox uses B (horizontal)
and G (vertical) channels for the displacements, with signed data (i.e.
0x00 means no displacement.) This is easy to see by modifying the game
files and experimenting. The boost dash 256x256 displacement map is in the
Media/Misc/Misc.dat file offset 0x60..0x40060. It's in BGRA byte order,
i.e. B is the first byte. Also to make observing the effect easier, the
bump map matrix can be modified. The game uses the 2x2 identity matrix
multiplied by 0.8, i.e. [0.8, 0.0; 0.0, 0.8]. The matrix can be found by
searching default.xbe for the bytes

CD CC 4C 3F  00 00 00 00  00 00 00 00  CD CC 4C 3F.

(At file offset 0x20BB48 for me.) As a curiosity, Xbox crashes if the
values in the matrix are set too large.

This commit produces the same result for JSRF as my Xbox v1.4 (and so does
also the current Xemu stable, I think.)