nv2a: Fix objects showing through closer surfaces
z_perspective is true implies w-buffering and then the w-coordinate stored
in the depth buffer should also be interpolated in a perspective-correct
way. We do this by calculating w and setting gl_FragDepth in the fragment
shader.

Since enabling polygon offset and setting values using glPolygonOffset
won't have any effect when manually setting gl_FragDepth for w-buffering,
we introduce the depthOffset variable to obtain similar behaviour (but the
glPolygonOffset factor-argument is currently not emulated.) (Note that
glPolygonOffset is OpenGL implementation-dependent and it might be good to
use depthOffset for z-buffering as well, but this is not done here and we
still use OpenGL/Vulkan zbias functionality.)

This also implements depth clipping and clamping in the fragment shader.
If triangles are clipped, the shadows of the small rocks in Halo 2 Beaver
Creek map can have flickering horizontal lines. The shadows are drawn on
the ground in another pass with the same models as for the ground, but for
some reason with depth clamping enabled. The flickering happens if Xemu
clips the ground triangles, but the exact same shadow triangles are depth
clamped, so there are small differences in the coordinates. The shadows
are drawn with depth function GL_EQUAL so there is no tolerance for any
differences. Clipping in the fragment shader solves the problem because
the ground and shadow triangles remain exactly the same regardless of
depth clipping/clamping. For some performance gain, it might be a good
idea to cull triangles by depth in the geometry shader, but this is not
implemented here.

Original commits:
f7d6253650a0652a177e70801b8b1f55dc9b97ae
8c5f4e8a115cc4ba28bbba12f10401886100459a
fb9b31143ca9e289762e01e0d49d96224a2a5d60

Superseded original commmits:
42fcd7cf71446821c625cd8e69b0d8ccfd7be97c
770343e5e342070beb46048045ca6c7c19864ac0
88a175aba1910c79f181520dace2a2e1320d2394
fc77e21bd986b56211cf1b2c5c8c9b62b233990f
e5255513b83572376bc888af8569bb305512cc0b