nv2a: Fix mipmapping with Intel integrated GPU, Vulkan and Mesa
This fixes a bug that occurs at least with Linux, Intel integrated UHD 770,
Mesa and Vulkan. The bug is that triangles are rendered with wireframes
when mipmapping is used. The conditional discards in the fragment shader
mean non-uniform flow control which can cause implicit derivatives (that
GPU computes to determine appropriate mipmap levels) to have undefined
results (as mentioned in GLSL documentation). This commit places the
conditional discards after sampling all textures so that sampling is done
in uniform flow control.

(Alternative fix that seems to work is to lower SPIR-V version from 1.6 to
1.5.)